requirements..txt
functions-framework==3.*
google-cloud-pubsub 
google-cloud
google-cloud-bigquery
google-cloud-core
